backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
asyncpg
pgvector
google-generativeai
python-dotenv
